+++ BetterReads/BetterReads/Model Controller/LibraryController.swift
@@ -212,181 +212,3 @@ class LibraryController {
         }.resume()
     }
 }
-
-struct UserShelf: Codable {
-    let shelfId: Int?
-    let userId: Int?
-    let shelfName: String?
-    let isPrivate: Bool?
-}
-/*
- [
-   {
-     "shelfId": 76,
-     "userId": 131,
-     "shelfName": "Philosophy",
-     "isPrivate": null
-   },
-   {
-     "shelfId": 77,
-     "userId": 131,
-     "shelfName": "Fire Stuff",
-     "isPrivate": null
-   },
-   {
-     "shelfId": 81,
-     "userId": 131,
-     "shelfName": "Sci-fi",
-     "isPrivate": null
-   }
- ]
- */
-
-/*
- struct Book: Codable {
-     let authors: [String]?
-     let categories: [String]?
-     let itemDescription: String?
-     let googleID: String? // wasn't optional
-     let isEbook: Bool? // wasn't optional
-     let language: String? // used to be Language and not optional
-     let pageCount: Int?
-     let publisher: String? // wasn't optional
-     let smallThumbnail: String? // wasn't optional
-     let textSnippet: String?
-     let thumbnail: String? // wasn't optional
-     let title: String? // wasn't optional
-     let webReaderLink: String? // wasn't optional
-     let averageRating: Double?
-     let isbn10, isbn13, publishedDate: String?
-
-     enum CodingKeys: String, CodingKey {
-         case authors, categories
-         case itemDescription = "description"
-         case googleID = "googleId"
-         case isEbook, language, pageCount, publisher, smallThumbnail,
-                textSnippet, thumbnail, title, webReaderLink,
-                averageRating, isbn10, isbn13, publishedDate
-     }
- }
- */
-
-struct UserBook: Codable {
-    let userBooksId: Int?
-    let bookId: Int?
-    let googleId: String?
-    let title: String?
-    let authors: String? // could this be an array?
-    let readingStatus: Int?
-    let favorite: Bool?
-    let categories: String?
-    let thumbnail: String?
-    let pageCount: Int?
-    let dateStarted: String? // DATETIME
-    let dateEnded: String? // DATETIME
-    let dateAdded: String? // DATETIME
-    let userRating: Double? // DECIMAL
-    // example response
-//    {
-//      "userBooksId": 365,
-//      "bookId": 315,
-//      "googleId": "OYtkbGl2j0sC",
-//      "title": "Fahrenheit 451",
-//      "authors": "Ray Bradbury",
-//      "readingStatus": 1,
-//      "favorite": false,
-//      "categories": "Fiction / Classics,Fiction / Science Fiction / General,Fiction / Media Tie-In",
-//      "thumbnail": "https://books.google.com/books/content?id=_api",
-//      "pageCount": 208,
-//      "dateStarted": null,
-//      "dateEnded": null,
-//      "dateAdded": "2020-05-11T18:50:21.337Z",
-//      "userRating": null
-//    }
-}
-
-// Example library response
-/*
- [
-   {
-     "userBooksId": 365,
-     "bookId": 315,
-     "googleId": "OYtkbGl2j0sC",
-     "title": "Fahrenheit 451",
-     "authors": "Ray Bradbury",
-     "readingStatus": 1,
-     "favorite": false,
-     "categories": "Fiction / Classics,Fiction / Science Fiction / General,Fiction / Media Tie-In",
-     "thumbnail": "https://books.google.com/books/content?id=OY1AR&source=gbs_api",
-     "pageCount": 208,
-     "dateStarted": null,
-     "dateEnded": null,
-     "dateAdded": "2020-05-11T18:50:21.337Z",
-     "userRating": null
-   },
-   {
-     "userBooksId": 364,
-     "bookId": 314,
-     "googleId": "kotPYEqx7kMC",
-     "title": "1984",
-     "authors": "George Orwell",
-     "readingStatus": 3,
-     "favorite": false,
-     "categories": "{\"Fiction\"}",
-     "thumbnail": "https://books.google.com/books/content?id=kocurl&source=gbs_api",
-     "pageCount": 648,
-     "dateStarted": null,
-     "dateEnded": null,
-     "dateAdded": "2020-05-11T18:44:50.051Z",
-     "userRating": null
-   },
-   {
-     "userBooksId": 363,
-     "bookId": 313,
-     "googleId": "CCiZnVG1j4cC",
-     "title": "Mortality",
-     "authors": "Christopher Hitchens",
-     "readingStatus": 3,
-     "favorite": false,
-     "categories": "{\"Biography & Autobiography\"}",
-     "thumbnail": "https://books.google.com/books/content?id=CC=curl&source=gbs_api",
-     "pageCount": 128,
-     "dateStarted": null,
-     "dateEnded": null,
-     "dateAdded": "2020-05-11T18:43:03.441Z",
-     "userRating": null
-   },
-   {
-     "userBooksId": 362,
-     "bookId": 312,
-     "googleId": "SGAZdjNfruYC",
-     "title": "Animal Farm",
-     "authors": "George Orwell",
-     "readingStatus": 2,
-     "favorite": false,
-     "categories": "{\"Fiction\"}",
-     "thumbnail": "https://books.google.com/books/content?id=SG&zoom=1&source=gbs_api",
-     "pageCount": 140,
-     "dateStarted": null,
-     "dateEnded": null,
-     "dateAdded": "2020-05-11T17:17:20.813Z",
-     "userRating": null
-   },
-   {
-     "userBooksId": 361,
-     "bookId": 311,
-     "googleId": "kcsqGna7fBIC",
-     "title": "Breaking Dawn",
-     "authors": "Stephenie Meyer",
-     "readingStatus": 1,
-     "favorite": false,
-     "categories": "{\"Young Adult Fiction\"}",
-     "thumbnail": "https://books.google.com/books/content?id=kc=curl&source=gbs_api",
-     "pageCount": 768,
-     "dateStarted": null,
-     "dateEnded": null,
-     "dateAdded": "2020-05-11T16:49:46.817Z",
-     "userRating": null
-   }
- ]
- */
