+++ BetterReads/BetterReads/Models/UserBookDetail.swift
@@ -8,26 +8,6 @@
 
 import Foundation
 
-/*
- struct UserBook: Codable {
-     let userBooksId: Int?
-     let bookId: Int?
-     let googleId: String?
-     let title: String?
-     let authors: String? // could this be an array?
-     let readingStatus: Int?
-     let favorite: Bool?
-     let categories: String?
-     let thumbnail: String?
-     let pageCount: Int?
-     let dateStarted: String? // DATETIME
-     let dateEnded: String? // DATETIME
-     let dateAdded: String? // DATETIME
-     let userRating: Double? // DECIMAL
- }
-
- */
-
 struct UserBookDetail: Codable {
   let bookId: Int?
   let userBooksId: Int?
